pyproject.toml
[project]
name = "resume-genie"
version = "0.1.0"
description = "Beautiful Streamlit app for resume review, job matching, cover letter generation, and AI career coaching."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-openai>=1.1.12",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.45.1",
]
